Now works with cloudsql
Should work with mysql in a dev env. Check settings.py for info. I'm unable to test because I'm on Windows with 64-bit Python, which has really, really awful support. Removing and installing the 32-bit version will almost certainly break my django install, so I'm committing as is and trying this later on Linux.