Djangofied everything again...
I've filled out the README. Please RTFM.

Some ugly UI had to go in here for the time being. Ryley, can you clean this up?
I had to add a title field to the ideas submission so we would have a title when
the idea gets promoted to be a real project. If you want this to work
differently, let me know.

Still needs a _lot_ of user checking and other auth. Most of this is needed to
prevent people from voting on a project more than once.